Adding logo and improving user feedback

+++ src/helper/helper.py
@@ -1,2 +1,2 @@
 def calculate_max_people(area):
-    return int(area/3)
+    return area/3
